pyproject.toml
[project]
name = "mongodb-session-manager"
version = "0.4.0"
description = "MongoDB session management for Strands Agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35.0",
    "fastapi>=0.116.1",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.13.2",
    "strands-agents>=1.12.0",
    "strands-agents-tools>=0.2.11",
    "uvicorn>=0.37.0",
    "uvloop>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.21.0",
]

[[project.authors]]
name = "Iñaki Guinea Beristain"
email = "iguinea@gmail.com"
keywords = ["strands", "agents", "mongodb", "session", "management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/iguinea/mongodb-session-manager"
Documentation = "https://github.com/iguinea/mongodb-session-manager/blob/main/README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mongodb_session_manager"]

